ml-service/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
prometheus-client
psycopg2-binary
pydantic
